requirements.txt
# Backend dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0
pypdf==3.17.0

# Text processing and chunking
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
tiktoken==0.5.2
nltk==3.8.1

# Embeddings and Vector Store
sentence-transformers==2.2.2
huggingface_hub==0.14.1
opensearch-py==2.4.0
faiss-cpu==1.13.2

# Google Gemini LLM
google-generativeai==0.3.2

# Frontend dependencies
streamlit==1.28.1
requests==2.31.0

# System monitoring and utilities
psutil==5.9.6
numpy>=1.25,<2
pandas==2.0.3
python-multipart==0.0.6
